created the events, commands and errors handler.
