+++ .env.example
@@ -5,3 +5,4 @@ BOT_TOKEN="z.y.x"
 UBFB_TOKEN="myubfbtoken"
 UBFB_PASSWORD="1234"
 DANBOT_TOKEN="mydanbottoken"
+BOT_PRIVATE_LOGS="ID_del_canal_donde_se_enviarán_los_logs"
