+++ eventos/shardConnect.js
@@ -1,0 +1,3 @@
+module.exports = (shardId, guilds) => {
+    console.log('Shard num' + shardId + ': LANZADO PARA ' + guilds.length + ' SERVIDORES.');
+};
